+++ modules/sdk/gradle-plugins-js-module-config-generator/CHANGELOG.markdown
@@ -137,6 +137,12 @@ to 1.2.1.
 - [LPS-73472]: Update the [Liferay Gradle Plugins Node] dependency to version
 2.3.0.
 
+## 2.1.16 - 2017-07-11
+
+### Changed
+- [LPS-73070]: Update the [Liferay Gradle Plugins Node] dependency to version
+2.4.0.
+
 [Liferay Gradle Plugins Node]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/gradle-plugins-node
 [LPS-66906]: https://issues.liferay.com/browse/LPS-66906
 [LPS-67023]: https://issues.liferay.com/browse/LPS-67023
@@ -156,4 +162,5 @@ to 1.2.1.
 [LPS-71826]: https://issues.liferay.com/browse/LPS-71826
 [LPS-72152]: https://issues.liferay.com/browse/LPS-72152
 [LPS-72340]: https://issues.liferay.com/browse/LPS-72340
+[LPS-73070]: https://issues.liferay.com/browse/LPS-73070
 [LPS-73472]: https://issues.liferay.com/browse/LPS-73472