+++ modules/sdk/gradle-plugins-js-transpiler/CHANGELOG.markdown
@@ -156,6 +156,12 @@ subdirectories created from the dependencies in the `jsCompile` configuration.
 - [LPS-73472]: Update the [Liferay Gradle Plugins Node] dependency to version
 2.3.0.
 
+## 2.3.16 - 2017-07-11
+
+### Changed
+- [LPS-73070]: Update the [Liferay Gradle Plugins Node] dependency to version
+2.4.0.
+
 [Liferay Gradle Plugins Node]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/gradle-plugins-node
 [LPS-66906]: https://issues.liferay.com/browse/LPS-66906
 [LPS-67023]: https://issues.liferay.com/browse/LPS-67023
@@ -178,4 +184,5 @@ subdirectories created from the dependencies in the `jsCompile` configuration.
 [LPS-72340]: https://issues.liferay.com/browse/LPS-72340
 [LPS-72723]: https://issues.liferay.com/browse/LPS-72723
 [LPS-72851]: https://issues.liferay.com/browse/LPS-72851
+[LPS-73070]: https://issues.liferay.com/browse/LPS-73070
 [LPS-73472]: https://issues.liferay.com/browse/LPS-73472